TripleX/requirements.txt
requests==2.32.3
scenedetect==0.6.5.1
beautifulsoup4==4.12.3
torch==2.3.1
torchvision==0.18.1
numpy~=1.26.4
pillow~=10.2.0
opencv-python~=4.8.0.76
tensorflow==2.18.0
gdown==5.2.0
openai~=0.28.1
tqdm~=4.66.3
praw~=7.8.1
prawcore~=2.4.0
python-dotenv~=1.0.0
google-generativeai
vertexai
google-cloud-storage
transformers
huggingface_hub
Pillow
